Moved pressure forces to PressureGenerator
