+++ rayon-core/src/main/java/dev/lazurite/rayon/core/impl/bullet/collision/space/generator/TerrainGenerator.java
@@ -1,7 +1,5 @@
 package dev.lazurite.rayon.core.impl.bullet.collision.space.generator;
 
-import com.jme3.math.Quaternion;
-import com.jme3.math.Vector3f;
 import dev.lazurite.rayon.core.impl.bullet.collision.body.ElementRigidBody;
 import dev.lazurite.rayon.core.impl.bullet.collision.body.TerrainObject;
 import dev.lazurite.rayon.core.impl.bullet.collision.body.shape.MinecraftShape;
@@ -142,53 +140,12 @@ public class TerrainGenerator {
         }
     }
 
-    public static void applyPressureForces(MinecraftSpace space) {
-        for (var rigidBody : space.getRigidBodiesByClass(ElementRigidBody.class)) {
-            var rigidBodyBox = Convert.toMinecraft(rigidBody.getCollisionShape().boundingBox(rigidBody.getPhysicsLocation(null), new Quaternion(), null));
-            rigidBodyBox = rigidBodyBox.contract(rigidBodyBox.getXLength() * 0.2, rigidBodyBox.getYLength() * 0.2, rigidBodyBox.getZLength() * 0.2);
-            var volume = 0.0f;
-
-            final var gravity = -1.0f * space.getGravity(null).y;
-            final var maxVolume = rigidBodyBox.getXLength() * rigidBodyBox.getYLength() * rigidBodyBox.getZLength();
-            final var fluidObjects = rigidBody.getTerrainObjects().values().stream()
-                    .filter(terrainObject -> terrainObject.getFluidState().isPresent()).toList();
-
-            final var relativePoints = new ArrayList<Vector3f>();
-            relativePoints.add(new Vector3f((float) rigidBodyBox.getXLength() * 0.5f, (float) rigidBodyBox.getYLength(), (float) (-1.0f * rigidBodyBox.getZLength() * 0.5f)));
-            relativePoints.add(new Vector3f((float) (-1.0f * rigidBodyBox.getXLength() * 0.5f), (float) rigidBodyBox.getYLength(), (float) (-1.0f * rigidBodyBox.getZLength() * 0.5f)));
-            relativePoints.add(new Vector3f((float) rigidBodyBox.getXLength() * 0.5f, (float) rigidBodyBox.getYLength(), (float) rigidBodyBox.getZLength() * 0.5f));
-            relativePoints.add(new Vector3f((float) (-1.0f * rigidBodyBox.getXLength() * 0.5f), (float) rigidBodyBox.getYLength(), (float) rigidBodyBox.getZLength() * 0.5f));
-
-            for (var fluidObject : fluidObjects) {
-                final var fluidBox = Convert.toMinecraft(fluidObject.getCollisionObject().boundingBox(null));
-
-                if (fluidBox.intersects(rigidBodyBox)) {
-                    final var intersection = fluidBox.intersection(rigidBodyBox);
-                    volume += intersection.getXLength() * intersection.getYLength() * intersection.getZLength();
-                }
-            }
-
-            final var force = new Vector3f(0.0f, (float) (gravity * 1000 * Math.min(volume, maxVolume)), 0.0f).multLocal(0.25f);
-
-            if (Float.isFinite(force.lengthSquared()) && force.lengthSquared() > 0.0f) {
-                rigidBody.applyDragForce(1000);
-
-                for (var point : relativePoints) {
-                    rigidBody.applyForce(force, point);
-                }
-            } else {
-                rigidBody.applyDragForce(1.2f);
-            }
-        }
-    }
-
     @Environment(EnvType.CLIENT)
-    public static Pattern tryGenerateShape(BlockView blockView, BlockPos blockPos, BlockState blockState) {
-        var transformation = new MatrixStack();
+    private static Pattern tryGenerateShape(BlockView blockView, BlockPos blockPos, BlockState blockState) {
+        final var blockEntity = blockView.getBlockEntity(blockPos);
+        final var transformation = new MatrixStack();
         transformation.scale(0.95f, 0.95f, 0.95f);
         transformation.translate(-0.5f, -0.5f, -0.5f);
-
-        var blockEntity = blockView.getBlockEntity(blockPos);
 
         try {
             if (blockEntity != null) {
